feat: kotlin multiplatform Keyple-less NFC reader library
This library wraps NFC reader functionality on iOS, Android and desktop targets, to allow raw communication with ISO-7816 NFC Cards.
In the context of Keyple less distributed protocol, this lib acts as the "driver" for the local NFC reader physically present on the machine (Android or iOS device, mac or PC...) and is used by the Keyple-less distributed client library.

Signed-off-by: Thierry Dassonville <tdassonville@gmail.com>